Version 1.0.2
- Beta Linux builds for X86, X86_64, ARM 32 (hard and soft), ARM 64.
- 4 example projects for Linux.
- Better folder structure for easier understanding.
- Fixed bug in SuperpoweredIOSAudioIO, where a recording-only category
returned 0 samplerate.

+++ Superpowered/SuperpoweredCPU.h
@@ -16,7 +16,7 @@
 */
 class SuperpoweredCPU {
 public:
-    static int sustainedPerformanceMode;
+    static long sustainedPerformanceMode;
 /**
  @brief Enable/disable the sustained performance mode. This method is thread-safe.
